Format NGP website webform exports for VAN upload
This script is used to take NGP website exports in .tsv format and
re-encode and reformat them for easier bulk upload to VAN. Save out
webform exports in .tsv format to one directory.  Create a second
directory for upload files. When prompted, enter the original directory
and then the target directory (minus the terminal \). All files in the
first directory will be converted and the resulting files will land in
the target directory.